+++ lib/userEvents/reportOnUserEvents.js
@@ -6,8 +6,8 @@ const EventTypes = require('./EventTypes');
 const renderUserEvent =  require('./renderHelpers').renderUserEvent;
 const renderIndex = require('./renderHelpers').renderIndex;
 
-const buildLinkToViewCreatedItem = require('./linkBuilders').buildLinkToViewCreatedItem;
-const buildLinkToViewPushEventCommits = require('./linkBuilders').buildLinkToViewPushEventCommits;
+const linkToViewPushEventCommits = require('./linkBuilders/linkToViewPushEventCommits');
+const linkToViewCreateEventItem = require('./linkBuilders/linkToViewCreateEventItem');
 
 const removeRefsHeads = (ref) => ref.substring('refs/heads/'.length);
 
@@ -51,10 +51,10 @@ const eventDescriptionBuilders = {
 
 const eventLinkBuilders = {
   [EventTypes.PushEvent]: [
-    buildLinkToViewPushEventCommits
+    linkToViewPushEventCommits
   ],
   [EventTypes.CreateEvent]: [
-    buildLinkToViewCreatedItem
+    linkToViewCreateEventItem
   ]
 };
